deleted unnecessary routes
